fix: move lifespan to _lifespan.py submodule, fix SIGTERM handler and integration test (#745)
- Move _autoskillit_lifespan from server/__init__.py to server/_lifespan.py
  to satisfy pure-facade and cross-package import arch rules
- Change SIGTERM handler from sys.exit(0) to raise KeyboardInterrupt for
  clean asyncio shutdown
- Fix integration test: use sys.executable -m, close stdin before SIGTERM
